added error for IOS reloading to stub, fixed IOS reloading for system menu, changed settings to make it smaller (THIS REV IS NOT COMPATABLE WITH OLDER REVS. but the installer cleans settings anyway) , made the HBC stub not load when its already loaded (rare cases) . also moved the stub loading to before settings get loaded. and to stop confusions a bit : it says 0.3c now eventho there is no public release of 0.3c

+++ trunk/priiloader/source/hacks.cpp
@@ -36,7 +36,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 std::vector<hack> hacks;
 u32 *states=NULL;
-extern u32 error;
+extern u8 error;
 
 unsigned int foff=0;
 char *GetLine( char *&astr, unsigned int len )
